--- histoprep/backend/__init__.py
@@ -1,5 +1,0 @@
-"""Slide reader backends."""
-
-from ._czi import *
-from ._openslide import *
-from ._pillow import *
